src/requirements.txt
psycopg[binary,pool]
fastapi
uvicorn
pyjwt
